Just lift the quotestring method from Parser.scala
