+++ querulous-core/src/main/scala/com/twitter/querulous/query/SqlQuery.scala
@@ -137,9 +137,38 @@ class SqlQuery(connection: Connection, val query: String, params: Any*) extends 
     if (annotations.isEmpty) {
       ""
     } else {
-      " /*~" + new JSONObject(annotations.toMap).toString() + "*/"
+      " /*~{" + annotations.map({ case (k,v) => "\"" + quoteString(k) + "\" : \"" +
+        quoteString(v) + "\"" }).mkString(", ") + "}*/"
     }
   }
+
+  /**
+   * TODO remove: this was lifted from Parser.scala in Scala 2.9, since the 2.8 version didn't escape
+   *
+   * This function can be used to properly quote Strings
+   * for JSON output.
+   */
+  def quoteString (s : String) : String =
+    s.map {
+      case '"'  => "\\\""
+      case '\\' => "\\\\"
+      case '/'  => "\\/"
+      case '\b' => "\\b"
+      case '\f' => "\\f"
+      case '\n' => "\\n"
+      case '\r' => "\\r"
+      case '\t' => "\\t"
+      /* We'll unicode escape any control characters. These include:
+       * 0x0 -> 0x1f  : ASCII Control (C0 Control Codes)
+       * 0x7f         : ASCII DELETE
+       * 0x80 -> 0x9f : C1 Control Codes
+       *
+       * Per RFC4627, section 2.5, we're not technically required to
+       * encode the C1 codes, but we do to be safe.
+       */
+      case c if ((c >= '\u0000' && c <= '\u001f') || (c >= '\u007f' && c <= '\u009f')) => "\\u%04x".format(c: Int)
+      case c => c
+    }.mkString
 
   private def buildStatement(connection: Connection, query: String, params: Any*) = {
     val statement = connection.prepareStatement(expandArrayParams(query + annotationsAsComment, params: _*))
